requirements.txt
# Core AI and ML dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
google-generativeai>=0.3.0

# Vector database and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
PyPDF2>=3.0.1
markdown>=3.5.1
python-docx>=0.8.11

# MCP framework
fastmcp>=0.1.0
mcp>=1.0.0

# Web framework for MCP server
fastapi>=0.104.1
uvicorn>=0.24.0

# Data processing and utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Text processing
nltk>=3.8.1
spacy>=3.7.0

# HTTP requests
requests>=2.31.0
httpx>=0.25.0

# JSON and configuration
pydantic>=2.5.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0 